* configure.ac: The correct option is --enable-plugin, not         --disable-plugin.         * backend/sound_handler_gst.cpp: Don't use realloc() on operator         new-allocated pointers. Do some trivial cleanup.         * backend/sound_handler_sdl.cpp: Trivially fix warnings.         * server/edit_text_character.cpp: Likewise.         * server/sound.cpp: Ensure that the sample rate is in the acceptable         range, to avoid invalid memory accesses.

+++ server/edit_text_character.cpp
@@ -176,7 +176,7 @@ bool edit_text_character::on_event(event_id id)
 					break;
 
 				case key::DELETEKEY:
-					if (s.size() > m_cursor)
+					if (s.size() > (unsigned int)m_cursor)
 					{
 						s.erase(m_cursor, 1);
 						set_text_value(s.c_str());
